Estrutura final aplicada:
Repositories com domínio collectionPointsQr:
handlers.ts
collectionPointsQrService.ts
questionsSnapshotRepository.ts
validation.ts
types.ts
Services com domínio iaAnalyze:
iaAnalyzeService.ts
iaAnalyzeGatewayClient.ts
iaAnalyzeErrors.ts
Imports atualizados nos endpoints:

EndpointsCollectionPointsQRCode.ts
EndpointsIAAnalyze.ts
