+++ backends/api-gateway/src/server/express/controllers/repositories/collectionPointsQr/handlers.ts
@@ -1,0 +1,9 @@
+export {
+  disableCatalogQrHandler,
+  disableQrHandler,
+  enableCatalogQrHandler,
+  enableQrHandler,
+  getQrCatalogHandler,
+  getQrStatusHandler,
+  upsertCatalogQuestionsHandler,
+} from './collectionPointsQrService.js';
