+++ backends/api-gateway/src/server/express/controllers/services/iaAnalyze/iaAnalyzeFeedbackRepository.ts
@@ -1,0 +1,298 @@
+import type { SupabaseClient } from '@supabase/supabase-js';
+import type {
+  IaAnalyzeFeedbackInput,
+} from 'lib/interfaces/contracts/ia-analyze/input.contract.js';
+import type {
+  IaAnalyzeScopeType,
+} from 'lib/interfaces/contracts/ia-analyze/scope.contract.js';
+import { IaAnalyzeServiceError } from './iaAnalyzeErrors.js';
+
+type SupabaseServerClient = SupabaseClient;
+
+type FeedbackForAnalysis = IaAnalyzeFeedbackInput;
+
+type RawFeedbackRow = {
+  id: string;
+  message: string;
+  rating: number | null;
+  created_at: string | null;
+  collection_points:
+    | {
+        id?: string | null;
+        name?: string | null;
+        type?: string | null;
+        identifier?: string | null;
+        catalog_item_id?: string | null;
+      }
+    | Array<{
+        id?: string | null;
+        name?: string | null;
+        type?: string | null;
+        identifier?: string | null;
+        catalog_item_id?: string | null;
+      }>
+    | null;
+};
+
+type RawCatalogItemRow = {
+  id: string;
+  name: string;
+  kind: string;
+  description: string | null;
+};
+
+type RawFeedbackQuestionAnswerRow = {
+  feedback_id: string;
+  question_id: string;
+  question_text_snapshot: string;
+  answer_value: 'PESSIMO' | 'RUIM' | 'MEDIANA' | 'BOA' | 'OTIMA';
+  answer_score: number;
+};
+
+type RawFeedbackSubquestionAnswerRow = {
+  feedback_id: string;
+  subquestion_id: string;
+  subquestion_text_snapshot: string;
+  answer_value: 'PESSIMO' | 'RUIM' | 'MEDIANA' | 'BOA' | 'OTIMA';
+  answer_score: number;
+};
+
+function normalizeScopeType(kind: string | null | undefined): IaAnalyzeScopeType {
+  const normalized = String(kind ?? '').toUpperCase();
+
+  if (normalized === 'PRODUCT') return 'PRODUCT';
+  if (normalized === 'SERVICE') return 'SERVICE';
+  if (normalized === 'DEPARTMENT') return 'DEPARTMENT';
+  return 'COMPANY';
+}
+
+function resolveCollectionPoint(
+  collectionPointRaw: RawFeedbackRow['collection_points'],
+) {
+  if (Array.isArray(collectionPointRaw)) {
+    return collectionPointRaw[0] ?? null;
+  }
+
+  return collectionPointRaw ?? null;
+}
+
+export async function fetchFeedbacksForAnalysis(params: {
+  supabase: SupabaseServerClient;
+  enterpriseId: string;
+  limit: number;
+}): Promise<FeedbackForAnalysis[]> {
+  const { supabase, enterpriseId, limit } = params;
+
+  const { data: feedbacks, error: feedbackError } = await supabase
+    .from('feedback')
+    .select(
+      `
+      id,
+      message,
+      rating,
+      created_at,
+      collection_points(
+        id,
+        name,
+        type,
+        identifier,
+        catalog_item_id
+      )
+    `,
+    )
+    .eq('enterprise_id', enterpriseId)
+    .order('created_at', { ascending: false })
+    .limit(limit);
+
+  if (feedbackError) {
+    throw new IaAnalyzeServiceError(
+      'Failed to fetch feedbacks for IA',
+      500,
+      'failed_to_fetch_feedbacks_for_ia',
+    );
+  }
+
+  const feedbackRows = (feedbacks ?? []) as RawFeedbackRow[];
+
+  const catalogItemIds = Array.from(
+    new Set(
+      feedbackRows
+        .map((feedback) => resolveCollectionPoint(feedback.collection_points)?.catalog_item_id)
+        .filter((id): id is string => typeof id === 'string' && id.length > 0),
+    ),
+  );
+
+  let catalogItemById = new Map<
+    string,
+    {
+      id: string;
+      name: string;
+      kind: 'PRODUCT' | 'SERVICE' | 'DEPARTMENT';
+      description: string | null;
+    }
+  >();
+
+  if (catalogItemIds.length > 0) {
+    const { data: catalogRows, error: catalogError } = await supabase
+      .from('catalog_items')
+      .select('id, name, kind, description')
+      .in('id', catalogItemIds);
+
+    if (catalogError) {
+      throw new IaAnalyzeServiceError(
+        'Failed to fetch catalog items for IA',
+        500,
+        'failed_to_fetch_catalog_items_for_ia',
+      );
+    }
+
+    catalogItemById = new Map(
+      ((catalogRows ?? []) as RawCatalogItemRow[])
+        .map((row) => {
+          const scopeType = normalizeScopeType(row.kind);
+          if (scopeType === 'COMPANY') return null;
+
+          return [
+            row.id,
+            {
+              id: row.id,
+              name: row.name,
+              kind: scopeType,
+              description: row.description ?? null,
+            },
+          ] as const;
+        })
+        .filter(
+          (
+            entry,
+          ): entry is readonly [
+            string,
+            {
+              id: string;
+              name: string;
+              kind: 'PRODUCT' | 'SERVICE' | 'DEPARTMENT';
+              description: string | null;
+            },
+          ] => Boolean(entry),
+        ),
+    );
+  }
+
+  const feedbackIds = feedbackRows
+    .map((feedback) => feedback.id)
+    .filter((id): id is string => typeof id === 'string' && id.length > 0);
+
+  const answersByFeedbackId = new Map<
+    string,
+    Array<{
+      question_id: string;
+      question_text_snapshot: string;
+      answer_value: 'PESSIMO' | 'RUIM' | 'MEDIANA' | 'BOA' | 'OTIMA';
+      answer_score: number;
+    }>
+  >();
+
+  const subanswersByFeedbackId = new Map<
+    string,
+    Array<{
+      subquestion_id: string;
+      subquestion_text_snapshot: string;
+      answer_value: 'PESSIMO' | 'RUIM' | 'MEDIANA' | 'BOA' | 'OTIMA';
+      answer_score: number;
+    }>
+  >();
+
+  if (feedbackIds.length > 0) {
+    const { data: answerRows, error: answersError } = await supabase
+      .from('feedback_question_answers')
+      .select(
+        'feedback_id, question_id, question_text_snapshot, answer_value, answer_score, created_at',
+      )
+      .in('feedback_id', feedbackIds)
+      .order('created_at', { ascending: true });
+
+    if (answersError) {
+      throw new IaAnalyzeServiceError(
+        'Failed to fetch feedback dynamic answers for IA',
+        500,
+        'failed_to_fetch_feedback_dynamic_answers_for_ia',
+      );
+    }
+
+    ((answerRows ?? []) as RawFeedbackQuestionAnswerRow[]).forEach((answer) => {
+      const current = answersByFeedbackId.get(answer.feedback_id) ?? [];
+
+      current.push({
+        question_id: answer.question_id,
+        question_text_snapshot: answer.question_text_snapshot,
+        answer_value: answer.answer_value,
+        answer_score: answer.answer_score,
+      });
+
+      answersByFeedbackId.set(answer.feedback_id, current);
+    });
+
+    const { data: subanswerRows, error: subanswersError } = await supabase
+      .from('feedback_subquestion_answers')
+      .select(
+        'feedback_id, subquestion_id, subquestion_text_snapshot, answer_value, answer_score, created_at',
+      )
+      .in('feedback_id', feedbackIds)
+      .order('created_at', { ascending: true });
+
+    if (subanswersError) {
+      throw new IaAnalyzeServiceError(
+        'Failed to fetch feedback dynamic subanswers for IA',
+        500,
+        'failed_to_fetch_feedback_dynamic_subanswers_for_ia',
+      );
+    }
+
+    ((subanswerRows ?? []) as RawFeedbackSubquestionAnswerRow[]).forEach(
+      (subanswer) => {
+        const current = subanswersByFeedbackId.get(subanswer.feedback_id) ?? [];
+
+        current.push({
+          subquestion_id: subanswer.subquestion_id,
+          subquestion_text_snapshot: subanswer.subquestion_text_snapshot,
+          answer_value: subanswer.answer_value,
+          answer_score: subanswer.answer_score,
+        });
+
+        subanswersByFeedbackId.set(subanswer.feedback_id, current);
+      },
+    );
+  }
+
+  return feedbackRows.map((feedback) => {
+    const collectionPoint = resolveCollectionPoint(feedback.collection_points);
+
+    const catalogItemId =
+      typeof collectionPoint?.catalog_item_id === 'string'
+        ? collectionPoint.catalog_item_id
+        : null;
+
+    const catalogItem = catalogItemId
+      ? (catalogItemById.get(catalogItemId) ?? null)
+      : null;
+
+    const scopeType = normalizeScopeType(catalogItem?.kind ?? null);
+
+    return {
+      id: feedback.id,
+      message: feedback.message,
+      rating: feedback.rating ?? null,
+      created_at: feedback.created_at ?? null,
+      scope_type: scopeType,
+      collection_point: {
+        id: collectionPoint?.id ?? null,
+        name: collectionPoint?.name ?? null,
+        type: collectionPoint?.type ?? null,
+        identifier: collectionPoint?.identifier ?? null,
+      },
+      catalog_item: catalogItem,
+      dynamic_answers: answersByFeedbackId.get(feedback.id) ?? [],
+      dynamic_subanswers: subanswersByFeedbackId.get(feedback.id) ?? [],
+    } satisfies FeedbackForAnalysis;
+  });
+}
